add mod_spidermonkey to msvc build, tweak mod_spidermonkey warnings, fix tts interface.
git-svn-id: http://svn.freeswitch.org/svn/freeswitch/trunk@686 d0543943-73ff-0310-b7d9-9358b9ac24b2

+++ src/include/switch_frame.h
@@ -53,7 +53,7 @@ struct switch_frame {
 	/*! the entire size of the buffer */
 	size_t buflen;
 	/*! the number of audio samples present (audio only) */
-	int samples;
+	size_t samples;
 	/*! the rate of the frame */
 	int rate;
 };
